pyproject.toml
[project]
name = "smnrpx"
version = "0.9.0"
requires-python = ">=3.12"
authors = [
  {name = "David Meyer", email = "dameyerdave@gmail.com"}
]
maintainers = [
  {name = "David Meyer", email = "dameyerdave@gmail.com"}
]
dependencies = [
  "jinja2>=3.1.6",
  "python-box[all]~=7.0",
  "pyyaml==6.0.3",
]

[formatting]
charset = "utf-8"

[tool.ruff]
line-length=100

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "PIE", "INT", "PLE", "DJ001", "DJ012", "C4"]

ignore = ["E501", "SIM102", "SIM114"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"
